+++ LoginFrame/BodyTea.cs
@@ -173,16 +173,23 @@ namespace LoginFrame
 
         private void button4_Click(object sender, EventArgs e)
         {
-            int userId = Convert.ToInt32(pageCtrl.dg.CurrentRow.Cells[0].Value);
-            if (userId > 0)
+            if (pageCtrl.dg.CurrentRow == null)
             {
-                DialogResult dr = MessageBox.Show("确定要删除'" + pageCtrl.dg.CurrentRow.Cells[1].Value + "'吗？", "确认删除", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
-                if (dr == DialogResult.OK)
-                {
-                    userService.deleteUser(userId);
-                    btnQueryClick();
-                }
+                MessageBox.Show("请选择一条记录");
             }
+            else
+            {
+                int userId = Convert.ToInt32(pageCtrl.dg.CurrentRow.Cells[0].Value);
+                if (userId > 0)
+                {
+                    DialogResult dr = MessageBox.Show("确定要删除'" + pageCtrl.dg.CurrentRow.Cells[1].Value + "'吗？", "确认删除", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
+                    if (dr == DialogResult.OK)
+                    {
+                        userService.deleteUser(userId);
+                        btnQueryClick();
+                    }
+                }
+            }            
         }
     }
 }
